Mv keydown listener to interactions; minor styling improvements
